requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
scipy>=1.7.0

# Data processing
datasets>=2.10.0
pandas>=1.3.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0

# Multilingual support
sentencepiece>=0.1.99
sacrebleu>=2.3.0

# Optional: For 8-bit model loading
bitsandbytes>=0.41.0
accelerate>=0.20.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
ruff>=0.0.260
mypy>=1.0.0
jupyter>=1.0.0
notebook>=6.5.0
matplotlib>=3.5.0
seaborn>=0.12.0